requirements.txt
discord.py==1.7.3
aiohttp==3.8.4